double_buffer

+++ cpp/test/test_lru_cache.cc
@@ -5,7 +5,7 @@ run(test_lru_cache) {
 
   for (int key = 1; key <= 10; key++) {
     assert_T(cache.put(key, 1));
-    assert_T(cache.size() == key);
+    assert_T(cache.size() == (size_t)key);
 
     auto iter = cache.find(key);
     assert_T(iter != cache.end());
@@ -13,13 +13,13 @@ run(test_lru_cache) {
   }
 
   assert_T(cache.put(11, 1));
-  assert_T(cache.size() == 10);
+  assert_T(cache.size() == 10u);
   assert_T(cache.find(1) == cache.end());
 
   assert_F(cache.put(3, 1));
   for (int key = 2; key <= 9; key++) {
     assert_T(cache.put(key + 10, 1));
-    assert_T(cache.size() == 10);
+    assert_T(cache.size() == 10u);
   }
   assert_T(cache.find(3) != cache.end());
 }
